requirements.txt
# LangGraph Log Analyzer Agent
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
PyGithub>=2.0.0
GitPython>=3.1.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.12.0
pydantic>=2.0.0
